requirements.txt
backcall==0.2.0
click==8.0.1
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==7.0.3
jupyter-core==4.8.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
pyzmq==22.3.0
six==1.16.0
SQLAlchemy==1.4.23
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
Werkzeug==2.0.1

scikit-learn~=1.0.1

xgboost==1.5.0